+++ flightlog.go
@@ -42,3 +42,39 @@ func (s *Service) GetFlightLogs(UserUUID uuid.UUID) (FlightLogs, dutil.Error) {
 	}
 	return res.Data.FlightLogs, nil
 }
+
+func (s *Service) GetFlightLog(userUUID, UUID uuid.UUID) (FlightLog, dutil.Error) {
+	// set path
+	s.serv.URL.Path = "/log/-"
+	// set query params
+	qs := url.Values{"userUUID": {userUUID.String()}, "UUID": {UUID.String()}}
+	s.serv.URL.RawQuery = qs.Encode()
+	// do request
+	r, e := s.serv.NewRequest("GET", s.serv.URL.String(), nil, nil)
+	if e != nil {
+		return FlightLog{}, e
+	}
+
+	// response structure
+	type Data struct {
+		FlightLog `json:"flightLog"`
+	}
+	res := struct {
+		Data   `json:"data"`
+		Errors dutil.Errors `json:"errors"`
+	}{}
+	// decode the response
+	_, e = s.serv.Decode(r, &res)
+	if e != nil {
+		return FlightLog{}, e
+	}
+
+	if r.StatusCode != 200 {
+		e := &dutil.Err{
+			Status: r.StatusCode,
+			Errors: res.Errors,
+		}
+		return FlightLog{}, e
+	}
+	return res.Data.FlightLog, nil
+}