+++ CHANGELOG.md
@@ -5,6 +5,11 @@ The format is based on [Keep a changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [0.2.0] - 2022-12-06
+### Added
+- Added the service method `GetFlightLog` which fetches a specific flight log
+related to a specific user for a specific flight log.
+
 ## [0.1.0] - 2022-12-05
 ### Added
 - The `AircraftType` type.